requirements.txt
flask
flask-socketio
influxdb-client
python-dotenv
eventlet
